Fix memory leak in Font_Handler::::draw_richtext (block_lines was not freed).
Fix memory leak in Font_Handler::join_sdl_surfaces. There was an early return that, if w or h was 0, prevented the execution to reach the call to SDL_FreeSurface. Make sure it is called. Also fix the arguments (make them const and pass the vector by reference instead of copying it).

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@1677 37b2a8de-5219-0410-9f54-a31bc463ab9c