Disable the make infrasturcture tool again.
With the current code it can not work, because it creates a headquarter, which is a conquering building, which adds an item to Editor_Game_Base::m_conquer_info, which is saved and loaded. But it is saved in Game_Game_Class_Data_Packet and Game_Game_Class_Data_Packet::Write takes an argument of type Game * and there is no Game object in the editor.

The solution to this problem is to not save/load m_conquer_info. It is redundant anyway, so it should most likely be eliminated completely. But that is for after the next release.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2078 37b2a8de-5219-0410-9f54-a31bc463ab9c