Bugfix: a race caused "ware: can't move from building X to Y (not a warehouse)"
When actProduce was execute while a production site's worker was currently
performing a fetch_from_flag, the worker would just drop the produced item
into the building, where it sits idle as a supply until somebody requests
it.
At that point, a Transfer will be created for the item which just sits in
the building, and since the building is no warehouse, the item cannot be
further transported, causing an exception.

Fixed the bug by delaying the actual production of an item until the worker
was safely inside the building without performing any tasks.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2374 37b2a8de-5219-0410-9f54-a31bc463ab9c