Improve compatibility with old savegames, see bug 1772626
When the program for an immovable is shorter than in the version of Widelands
that created the savegame, the program pointer is now reset if it's beyond
the end of program.

This repairs some older savegames and fixes the bug mentioned above, but it
will not guarantee a working game in all situations.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2397 37b2a8de-5219-0410-9f54-a31bc463ab9c