Fix trigger building option menu: * Since somewhere between revision 1964:1967 the dialog segfaulted when opened. Not any more. * In revision 1964 (and earlier) m_player and m_building was initialized to -1. Not any more. * Do not store copies of all building names in the dialog. * Get rid of magic numbers. * Eliminate all use of dynamic memory allocation in this code (41 uses of operator new eliminated). * When the player number is changed and the new player's tribe is different from the old player's tribe, check if the new tribe has a building with the same name as the building that was selected in the old tribe. If so, select the building with that name in the new tribe. Otherwise select the first building. * Make sure that the buttons are enabled when and only when they should be enabled. * Allow setting the radius to 0 (the smalles allowed radius was 1 before). * Remove the buttons to change the radius in steps of 100 and 10 (which made little sense).
Consider this a temporary fix. It does not mean that the interface is optional. Setting coordinates and radius should not require clicking buttons and keeping track of numeric values. The user should be able to press the mouse on a point on the map to select coordinates and drag from that point to select the radius. The trigger should be visualized on the map and also selectable so it can be dragged around, resized or its other options edited in a dialog. (See how a circle is edited in a vector drawing program (such as Inkscape).) Se feature request #1827733. It should also be possible to set the radius to infinite (for example by incrementing it to >= max(mapwidth, mapheight) / 2).

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2813 37b2a8de-5219-0410-9f54-a31bc463ab9c