Resolve inter-worker-type dependencies after all worker types have been loaded instead of during the simulation. This means that the worker type no longer needs to store a string with the name of the worker type that it can become. It just stores the index.
Introduce the type Ware_Index for ware (and worker) index. It is boxed for type safety and has a special value to indicate invalidity. But unlike the current hack where a lot of code assumes that this value is -1, the value is not used explicitly. It is instead referred to as Ware_Index::Null(). Ware_Index defines a convenient operator bool so that code like:
   int32_t const index = ...;
   if (index != -1)
      do_something;
   else
      handle_error;

can be replaced with:
   if (Ware_Index const index = ...)
      do_something;
   else
      handle_error;

The actual value of a ware index should not be used much, but when it is needed (for indexing arrays) it is available as Ware_Index::value(). When declaring a loop index for iterating over ware index it should have the type Ware_Index::value_t.

There are some temporary deprecated functions to interact with the many places in the code that still use int32_t for ware index.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2859 37b2a8de-5219-0410-9f54-a31bc463ab9c