Since leading_whitespace_checker now also detects trailing whitespace, misused tabs and some other errors, it is renamed to just whitespace_checker (although it does not detect all kinds of whitespace errors, such as missing operator padding). Let whitespace_checker do some more tests, that were detected with regexps before.
For the regexp tests, mark those that only detect things that whitespace_checker also detects, as redundant (but keep them so that they can still be used if whitespace_checker is missing for some reason). When running detect and whitespace_checker is found, skip regexps that are redundant with whitespace_checker and skip detect_spurious_indentation.py as well. This saves a lot of time (an example; from 3'47" to 1'44").

Fix some newly found errors in the C++ codebase.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2907 37b2a8de-5219-0410-9f54-a31bc463ab9c