Fix bug 1906380 (regression from r2945)
Notifications about removed buildings are sent on.

This bug was a typical case of refactoring code and not seeing a non-local
control statement (in this case, a return statement from within a for-loop)
while trying to cram everything into a single function instead of separating
the code into different functions.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2957 37b2a8de-5219-0410-9f54-a31bc463ab9c