Renderer: Rewrite the triangle renderer and fix texture coordinates.
The triangle renderer uses an edge-list based approach now, where triangles
are first clipped as polygons, then converted into lists of left- and right
edges and those lists are then rendered. This significantly cleans up the
code and simplifies the inner rendering loops conceptually.
It doesn't impact performance (if anything, it's about 1-2% faster, but that's
negligible).


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2982 37b2a8de-5219-0410-9f54-a31bc463ab9c