Fix an evil bug source by making constructor explicit.
This is also a nice reminder of why I'm not overly fond of being
reference-crazy - because reference-crazy people often introduce references
in places where we really *CANNOT* treat the passed object as plain old data,
because we're taking its address and other stuff.

This particular madness came from the fact that DirAnimations are passed
by reference to Bob::start_task_forcemove(), which is just plain insane,
because start_task_forcemove() does care about the lifetime of the given
object. </rant>


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3010 37b2a8de-5219-0410-9f54-a31bc463ab9c