Simplify Bob tasks.
The central change here is that signals no longer receive special treatment
via signal handlers; the update() routine is now responsible for handling
them. This should simplify the semantics of when update() is called
significantly.

Unfortunately, this might introduce regressions, but I'm hopeful that it makes
it easier to maintain this type of code.

Also, this happens to fix bug 1910405.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3012 37b2a8de-5219-0410-9f54-a31bc463ab9c