Extend the power of expression in Requirements.
Building requires the Boost libraries now.

Realize that the training sites don't really want "and"-type requirements,
but "or"-type requirements. In other words, any soldier is fine as long as
he can be upgraded in at least one of the categories. The "and"-based logic
was really not correct.

This also fixes bug 1910232.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3016 37b2a8de-5219-0410-9f54-a31bc463ab9c