Remove the ability to load a game during a game.
This code path has a number of problems that are related to the unnice
relationship between Interactive_Player, Game, and whoever started it all.
Besides, it was poorly tested and therefore buggy. While it's nice to have
this ability, it doesn't seem to be used too much, and it's nasty to
maintain, so let's just get rid of it.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3029 37b2a8de-5219-0410-9f54-a31bc463ab9c