Fix desync: Tie economy rebalance commands to flag instead of pointer.
The pointer-based system was flawed, because during the time a command was
waiting in the Cmd_Queue, an Economy could be deleted, and - depending on the
system - a new Economy could be created at the same pointer address.

I don't like introducing such a subtle change this late before a release,
but the alternative (and arguably cleaner) fix to derive Economy from
Map_Object would be a lot more intrusive.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3147 37b2a8de-5219-0410-9f54-a31bc463ab9c