Use the standard section/key/value syntax instead of a special-purpouse syntax for productionsite types' working positions. This means that instead of:    [global]    worker=worker_type_a,worker_type_a,worker_type_a,worker_type_b it is now:    [working positions]    worker_type_a=3    worker_type_b=1
In Event_Building:
* Allow filling builds with things that they would normally request (wares, workers, soldiers). Previously it was only possible to fill a productionsite with all workers. Now it is possible to select exactly how many workers of each type that should be created (for the others that are needed, requests are created as usual). It is also possible to fill warehouses with any amount of wares, workers and soldiers. Still missing is code to fill the input queues of productionsites with wares and soldier controls (militarysites and traningsites) with soldiers.

Do not implicityly create headquarters for players in scenarios. Instead, add events to do that explicitly in each scenario. This change makes it possible to set exactly when and with what content the headquarter (or other constellation of buildings) is created. It is for example possible to have a (computer controlled) player that does not start with anything, but when the (human) player explores a certain point, the other player is created with a warehouse and some other buildings.

In the worker list window of productionsites; if a worker is missing, show "(vacant)" if the request for it is open and "(coming)" if the request is in transfer.

Implement the part of feature request #1797118 (get rid of line numbers in programs) that applies to immovable programs.

Implement feature request #2138397 (tree lifecycle dynamics):
* Trees have natural reproduction. Instead of just growing up and then idling forever, they seed new trees and eventually die. Dead trees disappear after a while.
* Increase the working radius of woodcutters/lumberjacks from 8 to 10. This makes it easier to cover large areas with a few of them. This is now recommended to keep an area clear of trees.

Implement feature request #2294333 (automatic recursive road removal):
* When a player holds down Ctrl when ordering (or confirming, if applicable) the removal of flag or road, do a recursive removal. If a road removal command is given with the recursive option, all roads between the start and end flags are removed. Then those flags that become dead ends are removed recursively. A flag is a dead end if it does not have a building and roads to at most 1 other flag.

When reading packet versions from conf-files, use get_positive instead of get_int, since packet versions are (and always should be) positive integers.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3591 37b2a8de-5219-0410-9f54-a31bc463ab9c