In Tribe_Descr; get rid of load_warehouse_with_start_wares, m_startwares, m_startworkers and m_startsoldiers.
Do not hardcode the creation of a building of the type "headquarters" at each player's starting location in non-scenario games. Instead each tribe has a set of initializations. Each initialization has a name, a descname and a set of events that are executed for the player at his starting position when the game begins.

Currently each tribe must have an initialization called "headquarters_medium". This initialization contains an event that creates a headquarter with the wares/workers/soldiers that were previously defined in the sections startwares/startworkers/startsoldiers. So the default behaviour is unchanged.

It is possible to define other initializations that can get the players started. One that I tried, not involving a headquarter, was to create a regular warehouse, give vision with radius 13 and conquer with radius 9. Someone might be interested in creating an initialization called "headquarters_minimal" with only a minimal set of wares that are needed to get started. Another initialization would be "headquarters_plenty" with lots of resources for a really fast start.

Of course the hardcoded use of "headquarters_medium" should be eliminated. It should be possible to set the initialization for each player in the playerdescgroup, just like his tribe. Usually all players should have the same initialization (and an initialization in one tribe should be balanced against initializations with the same name in other tribes). But sometimes an experienced player may get an initialization like "headquarters_minimal" against a beginner that gets "headquarters_medium".

In the future it should be possible to have initializations with several different buildings, for example a fortress + a regular warehouse instead of headquarters. There should also be an initialization that includes some basic buildings, like toolsmith, lime-kiln (barbarians) and stonemason (empire) and some trainingsites. But that requires implementing some fancy logic for building placements since there can only be one building at the starting location.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3616 37b2a8de-5219-0410-9f54-a31bc463ab9c