Add some debug output to hunt down #2868683.
Here is what might happen.

The worker 3755 walks along the road 6330. He has a task transfer. On top of that, he has a task movepath, with the path of the road. Bob::movepath_update calls start_task_move (from (49, 143) in direction 4 (sw) towards (49, 144)), which calls push_task, which schedules an act after 10ms and then returns.

Then the road is split by a flag that is placed at (49, 143). The road sees that the worker is on a location that is still on the path of the shrunken road ((49, 141), (50, 142), (49, 143)), so the worker is not reassigned to the new road. The road sends the signal 'road' to the worker.

When the 10ms has passed, the worker acts. Bob::move_update is called. It sees that state.diranims is true, so it calls start_walk, which updates the worker's position to the next node (49, 144). start_walk returns the time that it takes to walk the edge. move_update uses that duration to scheduale a new act.

When the worker acts again, it sees that the walkend time is <= the gametime and knows that it is done. It calls pop_task and returns. pop_task schedules and act after 10 ms.

When the 10 ms has passed, the worker acts. Bob::movepath_update is called. It sees that the worker has a signal and therefore calls pop_task. pop_task schedules and act after 10 ms.

When the 10 ms has passed, the worker acts. Worker::transfer_update is called. It sees that the signal is 'road' or 'fail' and ignores it. It sees that the worker's location is not a Building, not a Flag, but a Road. This is the shrunken road that the worker is no longer on!. The function calls start_task_movepath on the worker with the road's path. It fails because the worker is no longer on that path.


The reason that this bug does not happen more frequently is that the road must be split during the 10ms after the move task has been pushed and before the worker acts and calls move_update, which calls start_walk, which updates the worker's position. If the road is split after that, the split code will see that the worker is on the second part of the road, and reassign it.

So how should this bug be fixed? Would it work to remove that 10ms time window, by calling schedule_act with 0 delay? There seems to be several such short arbitrarily sized time windows in various places. After seeing this bug, it seems like a good idea to get rid of them. Is anyone aware of any proble with 0 delays in the command queue? Is it so, that if the act is scheduled with 0 delay, a playercommand, such as a road split, can not come between?

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@4594 37b2a8de-5219-0410-9f54-a31bc463ab9c