merge fs-optimisation: Add the results from FS_Tokenize directly into the node list.
FS_Tokenize was returning a list of file path nodes that were
being inserted into the start of the path nodes list being
canonicalized. This is inefficient, as it produces copies of the
nodes being tokenised.

The algorithm has been changed to take an output iterator, and an
inserter is being used to add the nodes directly into the
components list.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@5008 37b2a8de-5219-0410-9f54-a31bc463ab9c