merge fs-optimisation: On posix systems, don't generate a trailing '/'.
When taking the list<string> and converting it to a path string,
the previous algorithm (and what is still being done on the
Windows version) built it up as 'path/' then erasing the last '/'
character.

Since a path will always start with either './' or '/', the
algorithm can be changed to build the path up as '/path' instead,
removing the need to erase the trailing '/' character.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@5009 37b2a8de-5219-0410-9f54-a31bc463ab9c