global persistence now works with some bugs...
I also added a simple reference holder for a coroutine. This guarantees
that the thread is not garbage collected. This is very simple and just
keeps a key, value combination in the lua registry.

The following problems remain:
Problem 1) Pickling a coroutine always throws: pickling a c function
Problem 2) not the global dictionary is pickled, instead blub is. This
can be fixed by copying all entries from the loaded to the global
dictionary.
Problem 3) the code is a total mess currently.