Improve fighting-related game performance (bug #536211)
This patch tweaks the persist values that are used for soldier path finding,
and makes soldiers slightly less aggressive about re-finding paths, relying
instead on the "wakeup" signals sent by sendSpaceSignals().

In addition, the self-check code of the heap used in path finding is
deactivated, and the performance of Soldier::checkNodeBlocked() slightly
improved.

Overall, this causes much less CPU time to be spent on fights. However,
a large number of active soldiers can of course still cause heavy, and
particularly irregular CPU loads.