Fix a subtle bug when roads are removed and built again
When a road attached to a flag was removed while its carrier was on its way
to pick up an item, then depending on the precise circumstances, this item
was sometimes not reset as waiting for a carrier.

As a result, when the exact same road was then later rebuilt, the item could
get stuck "forever".