Fix bug #536373: Race between "transfer" and "cancel" signals
If a gowarehouse task received a transfer (and thus a "transfer" signal), but
the transfer is almost immediately cancelled before the gowarehouse task is
replaced by a transfer task, the transfer wasn't cleaned up properly. Fix this
problem.

Additionally, don't fail a transfer immediately just because we've been
blocked by a battle (this avoid crazy ping-pong generation and cancellation
of transfers, and so should also help performance during battle).