Bug 536464: Pause button
The Pause button pauses the game. Pressing it again returns to old speed.
Disabled in network games. The code that displays the game speed status shows
now, for example, "PAUSE (3x)", which works nicely and thus does not need to be
adjusted.

The P button is already used in the editor and thus not assigned pause.