Soldiers always move out of buildings to battle
In certain situations, a defending soldier can be assigned a battle while
returning into their building; it is also conceivable that a player builds
a building on top of a waiting soldier, thus preventing the other soldier
from reaching them.

To fix this, a soldier will always move out of a building it is in first,
unless we're already on the same field as the opponent.

This fixes bug #612348.