WareInstance no longer accepts buildings as locations
It is important that the economy of a WareInstance automatically gets
updated when the economy of its location changes. However, buildings
do not take care of updating associated WareInstances. This is normally
not noticed because the time window in which items may be in such a
state is extremely short.

This change should make the code more robust: An item that arrives at
a building is immediately forwarded to whatever the next step is,
and items now can only be located on Flags, Workers, PortDocks or Ships,
all of which take care of updating an item's economy properly,
and remove the item properly when they get destroyed.