Fix some of the new warnings caused by the previous merge
In many cases, the treatment of the default case simply needed to be moved
explicitly into the switch statement that came before.

In those cases where a switch in a potentially tighter inner loop was
not supposed to go into a default label, a default label with an assert
statement was added.