GameRendererGl4: use shaders to off-load more computation to the GPU
This new terrain rendering path draws base textures + dithers in a single
pass. The CPU no longer has to explicitly produce a list of triangles.
Instead, there's a template 'patch', and each frame a bunch of instances
of this patch are rendered with appropriate per-instance data for the
offsets.

This is incomplete; some of the TODOs:
- proper switching between the new and old rendering paths
- road rendering hasn't been implemented yet
 -> hard to avoid producing triangles for this on the CPU
 -> very simple vertex structures should suffice since the GPU already
    has all the height and brightness data, and simply vertex structures
    can be generated quickly on the CPU
- player perspective (fog of war) hasn't been implemented yet
- a lot of data is still re-uploaded every frame, this is unnecessary
  and a lot of CPU can still be saved
- minimap rendering could re-use this infrastructure