GameRendererGl4: implement road rendering
This is basically a mix-up of the GL2 path for roads and the GL4 path for
blits: the CPU is responsible for collecting the road segments to be
rendered, but data is uploaded per-road into a buffer. The vertex shader
builds all the per-vertex data based on the per-road data and the vertex ID.

As a future cleanup, shader code should be shared between terrain and road
rendering.