Fix time display in General Statistics
Fixes #3457
This fixes the problem that the time on the X axis in the General Statistics window is shown incorrectly if the slider is set to "game" and the game length is between 32 and 34 hours, or between 36 and 38 hours.
The fix decreases how_many_ticks until it is a divisor of max_x, because the label drawing code assumes it is the case and doesn't work properly if it isn't.