Make use of stages
We use builds stages now. First we run a code check and test the documentation. If nothing fails we the can go on and start compiling. This saves roughly 2 minutes on the debug builds which should help us with timeouts. Furthermore, we don't waste time when the code isn't clean.