revamp Travis-CI builds (#3508)
Re. #3493

Travis-CI builds are completely rewritten and don't require sudo anymore.

macOS: We build on 10.14(clang 7), 10.13(clang 5) & 10.12(clang3.9)
Linux: We build on 18.04 (clang 7 & gcc 7.4), 16.04 (gcc 5.4) & 14.04 (clang 3.4 & gcc 4.8)

On macOS all builds use ccache on Linux only the gcc Release builds. Since clang compiles
faster it is not as big of a problem and jobs are within the travis time-limit.

Regression suite is run on Linux jobs only if the compile-time didn't exceed 33 mins, to
avoid hitting the time-limit of 50 mins while executing the tests. This mainly affects
gcc debug builds.

For xvfb we utilize the new service stanza - only available on 18.04 and 16.04, though.
For 14.04 we use the old fashioned way.

Only pushes to master and PR trigger builds.

We use builds stages now. First we run a code check and test the documentation. If nothing fails we the can go on and start compiling.