Fix wares stuck on flag after loading (#3712)
Legal values for Carrier::promised_pickup_to_ are -1 to 1. Previously values under -1 were used.
This compatibility code was supposed to change them to -1 when loading saved games with the old values.
Instead it changed values *over* -1 which could cause carriers to forget they acked a ware, and never go to pick it up.