Prettier stdout files when regression_test is run with python 3
Previously, the output was printed as raw byte strings and the output
started like this:

b'This is Widelands Version r24504[28e2244@reg-test-output] (Debug)\n'b'Adding home directory: ...

i.e., everything on one line, with b' and ' enclosing every output from
widelands. Note that testing was still correct using python 3, but
debugging the output files was not easy.