Simplifying the Fellowships map, for better playing experience. (#3792)
This modifies the "Fellowships" map.

Expanding to the southern island requires many small military buildings, and one middle sized one. One has to know which location requires the middle sized building. A casual player first makes a small building, dismantles it and build the larger one. This is annoying. Playing should be pleasant not annoying. Requiring the middle sized one was a bad design choice.

This modifies the map slightly so, all buildings can be small while expanding to south island. No other modifications, unless I was careless with mouse.