Allow preserving wares left in buildings when dismantling or enhancing.
Any wares left in the building will, if desired, be dropped out by the builder immediately upon his arrival at the PartiallyFinishedBuilding.
Whether this is desired can be set via a checkbox in the confirmation dialogue. This option is checked by default.
Ctrl-clicking the dismantle or enhance button will discard any wares left in the building.
The AI will always preserve wares when enhancing or dismantling sites.