Fix scenario details on scenario select screen
The scenario description doesn't fit in the box at 800x600 resolution.
It's because scenario_details_ doesn't take into account
scenario_difficulty_ and scenario_difficulty_header_, which were added
in Build 21.
We need to include the height of these two elements and padding (2x)
in our calculations.

Related to widelands/widelands#3985