Target workers to immovables (#4209)
Allow marking/unmarking an immovable for ASAP removal by a worker. The button for this (in the fieldaction window) is only shown if there actually is a building within range that gathers either this immovable, or one of its future types (e.g. you can already mark a young tree for removal and it will be removed as soon as it's old enough).

Manually marking immovables like this overrides the worker's preference to walk short distances. If there is a marked immovable around, the worker will always pick one of those. If a worker can choose between multiple marked immovables, choice is random with no distance bias.