Fixing not being able to load multiplayer games. (#4402)
Fixing bug #4367 "Client can't join saved network game". The visible bug was that the client received a network packet about an unknown player.
The problem was that the UI gets the update notification for a map change before the clients receive the message about the map change (and consequently the number of existing player positions). As a result to the notification the UI sends player updates (about unknown players) to the clients, leading to the bug.
This branch fixes the problem by sending the UI notification after the map change has been send to the client.