Use two rows if there are less than 120 pixels available for six buttons.
Problem is that one button takes space at least 20 pixels. Six buttons do
not fit nicely into the Minimap if its width is smaller than 120 pixels.
This happens with maps that are narrower than 120 pixels with or without
zoom.

The commit fixes the problem and the bug #4528.

An interesting map size is 64x192px. Its width is narrower than 120px
without zoom and over it with 2X zoom when 800x600px window resolution
is used.