No custom starting positions for Discovery (#4765)
Disable the Custom Starting Positions checkbox if none of the players have a starting condition for which it has any effect. If some players use Discovery and some use HQ, the checkbox is enabled but the starting position selection is skipped for those players for whom it makes no sense.