Update defaultai_warfare.cc
It does not make sense to check the amount of mines to determine whether a trainingsite should be enhanced or not. 

Only the Empire has expandable training buildings (Arena -> Colosseum), no other tribe. So I'm guessing the author of this line wanted to make sure the Empire had enough marble and gold before an arena was upgraded to a colosseum. On the other hand the Colosseum used no weapons and therefore no metals for training. 

But for all other tribe developers it's an unnecessary limitation. For example, the Amazons could not get upgradeable training pages because they do not have iron mines and therefore will never meet this requirement.

If it doesn't cause any side effects, this condition should removed permanently.