Fix race in use of logic_thread_locked.
Need to use compare_exchange instead of testing and setting the value
separately. This fixes a race when ending a game and UI thread sets
kEndingRequested but logic thread ends up overwriting it with kFree/kLocked. In
such case UI is stuck showing 'waiting for game to end' but logic never stops.