merge_and_push_translations.sh generates files on matching changes
Files (authors, appdata, statistics) is generated for a commit of
translation and for a commit of catalogs.

Desired effects:
When translations are edited, only commit "translation" is created.
When translation sources change, only commit "catalog" is created.
When both change, both commits are created.
When a file was changed (on release, ...) and no translation changes,
  the commit "catalog" is created.
No commit triggers the other one.