extend documentation about attributes of units
The documentation for lua api only mentions has_attribute() for ImmovableDescription.
But it exists for all MapObjectDescriptions. It will return false for all types
except of ImmovableDescription and critters.
The documentation about writing units tells attributes can be set, but does not
mention they do not work for all units.

improves #4744