lua doc, correct example for getting the interactive player
In data/scripting/training_wheel/ there is different code to get the interactive player.
Experiment showed the example in the doc is wrong.
This only matters if any player was deselected which is listed before the interactive one.
(Was "closed" when selecting the players, or "shared in" for network games.)