Cancel battles only when players negotiate real truce
Forbidden attack only for one side happens because of hostage in campaign atl02.wmf,
and may happen in addon scenarios (or scripts).
In such a case the soldiers were blocked because one wanted to attack continously,
but the other rejected. Now defending when attacked is possible again, as in v1.1.

completes #5665