smaller map for all_tribes.wmf
Just have two positions for the players. As there is only one placed at a time, this is enough.
Reduce the vertical space as well, it is unused.

This speeds up loading of the map (factor 2).
The first game time (map loaded till movement of animals) happens earlier (factor 4).
The same for first movement of the workers after having placed the buildings (factor 4).
The game time can run much faster now (factor 4).
As this all happens for each tribe, it sums up.

Of course this updates the version of the map.