regression_test.py uses other error code for ASan memory leaks
This makes it easier to distinguish those failures, for people and code.

The original error code of the program is overwritten when a memory leak is detected,
which is not a big deal because output is checked anyway.
When any other error happens, it is unlikely that it is the same error code as from a memory leak.