remove 2/3 of the ducks from all_tribes.wmf
They do not help, but slow down the game.
This allows faster animation speed. (On my machine, run of 1st test was 2:45 and is 2:3 now.)
And animation (movement of the critters) starts a bit earlyer.

Do it in script as I can do this easier than by editing the map.
Do it before the 1st sleep, else it is too late for the 1st movement.