Embed a copy of libmd with the option whether to build with that or the system-wide libmd (if any), like for minizip.
Linux and others with a decent system installation can keep using that. The Windows and Mac CI can use the embedded sources.
By default, we detect whether a system-wide installation exists, and try to use that if present and fall back to the embedded sources otherwise.
Use `-DOPTION_FORCE_EMBEDDED_LIBMD=ON` to force using the embedded sources.

Also: libmd comes with a SHA1 implementation that we now use as well, so we can drop the existing embedded third-party SHA1 implementation.
Added a pair of SHA1 testcases to the unit tests as well.