fix assertion thrown by ship expedition
Fix assertion error triggered in Widelands::ShippingSchedule::ship_added(...).

An assertion error was triggered when a ships expedition was cancelled after it was sent to a destination.
This happened when the ship was on the way or already had reached the goal.
The behaviour of a non development build remains the same.

No test, as sending a ship to a destination while doing an expedition is not implemented in scripting API.