find_unused_includes.py caches results of one function
Like this the files do not have to be read and analysed repeatedly.

On my system, the time for the this util goes from 5 sec to around 1 sec.
This does not help a lot on the build system,
but it helps if you run the script locally before a commit.