build_deps.py pre-checks if line can match any of the patterns
Checking for `someText in longText` is faster than matching a regexp which contains someText.

On my system, the time for the this util goes from 2.5 sec to around 0.4 sec.
This does not help a lot on the build system,
but it helps if you run the script locally before a commit.