Attempt to fix MacOS builds not working on anything but the latest Mac OS
build_app.sh contained a small logic bug: The problem was that when building on newer macOS runners (15.x), the script was incorrectly setting the deployment target to match the runner's OS version instead of keeping it at the intended minimum version.

Before:
- Script set deployment target to 12.3
- When macOS 12.3 SDK wasn't found on GitHub Actions runners, it reset
  deployment target to current OS version (15.x)
- Result: binaries required macOS 15.7+

After:
- Deployment target is set to 11.0 and stays at 11.0
- Even when using a macOS 15 SDK (because that's what's available), the deployment target remains 11.0
- Result: binaries will work on macOS 11.0 and newer

This is replicating the manual fix that worked for me, i.e.  `vtool -set-build-version macos 11.0 15.5`. This changes the deployment target to 11.0, while keeping the SDK at 15.5.

Side Effects:
- DMG filenames will now include "11.0" instead of "12.3" (from line 64: widelands_${OSX_MIN_VERSION}_${WLVERSION}.dmg)
- Binaries will be compatible with macOS 11.0 Big Sur and newer (released in 2020)